filter by date range option on dialog implemented
